Add ViewModel for Dice
